refactor: create popover shared component, and make LetterCasePicker use it.
